+++ src-tauri/src/command.rs
@@ -4,4 +4,5 @@ pub mod entity;
 mod error;
 pub mod external_recipe;
 pub mod ocr;
+pub mod unit_conversion;
 pub mod unit_list;