+++ src-tauri/src/command/entity.rs
@@ -4,3 +4,4 @@ pub mod recipe_file;
 pub mod recipe_ingredient;
 pub mod recipe_ingredient_draft;
 pub mod recipe_step;
+pub mod unit_name;