+++ src-tauri/src/entity_crud.rs
@@ -18,6 +18,7 @@ pub mod recipe_file;
 pub mod recipe_ingredient;
 pub mod recipe_ingredient_draft;
 pub mod recipe_step;
+pub mod unit_name;
 
 /// This struct represents just getting the id column of a database table.
 ///