+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Types of changes:
 - Implement recipe files (image/video/audio) with optical character recognition
 - Add field "quality" to recipe ingredients
 - Implement getting external recipes
+- Implement unit conversion
 
 ### Changed